+++ src/promptflow/promptflow/_sdk/_submitter/test_submitter.py
@@ -14,8 +14,8 @@ from promptflow._internal import ConnectionManager
 from promptflow._sdk._constants import LOGGER_NAME, PROMPT_FLOW_DIR_NAME
 from promptflow._sdk._utils import dump_flow_result, parse_variant
 from promptflow._sdk.entities._flow import Flow, FlowContext
+from promptflow._sdk.operations._flow_context_resolver import FlowContextResolver
 from promptflow._sdk.operations._local_storage_operations import LoggerOperations
-from promptflow._sdk.operations._run_submitter import SubmitterHelper, variant_overwrite_context
 from promptflow._utils.context_utils import _change_working_dir
 from promptflow._utils.exception_utils import ErrorResponse
 from promptflow._utils.multimedia_utils import persist_multimedia_data
@@ -23,6 +23,8 @@ from promptflow.contracts.flow import Flow as ExecutableFlow
 from promptflow.contracts.run_info import Status
 from promptflow.exceptions import UserErrorException
 from promptflow.storage._run_storage import DefaultRunStorage
+
+from .utils import SubmitterHelper, variant_overwrite_context
 
 logger = logging.getLogger(LOGGER_NAME)
 
@@ -51,7 +53,7 @@ class TestSubmitter:
             tuning_node, node_variant = parse_variant(self.flow_context.variant)
         else:
             tuning_node, node_variant = None, None
-        self.flow_context._resolve_connections()
+
         with variant_overwrite_context(
             flow_path=self._origin_flow.code,
             tuning_node=tuning_node,
@@ -70,7 +72,9 @@ class TestSubmitter:
                 self._tuning_node = None
                 self._node_variant = None
 
-    def resolve_data(self, node_name: str = None, inputs: dict = None, chat_history_name: str = None):
+    def resolve_data(
+        self, node_name: str = None, inputs: dict = None, chat_history_name: str = None, dataplane_flow=None
+    ):
         """
         Resolve input to flow/node test inputs.
         Raise user error when missing required inputs. And log warning when unknown inputs appeared.
@@ -86,17 +90,23 @@ class TestSubmitter:
         """
         from promptflow.contracts.flow import InputValueType
 
+        # TODO: only store dataplane flow in context resolver
+        dataplane_flow = dataplane_flow or self.dataplane_flow
         inputs = (inputs or {}).copy()
         flow_inputs, dependency_nodes_outputs, merged_inputs = {}, {}, {}
         missing_inputs = []
         # Using default value of inputs as flow input
         if node_name:
-            node = next(filter(lambda item: item.name == node_name, self.dataplane_flow.nodes), None)
+            node = next(filter(lambda item: item.name == node_name, dataplane_flow.nodes), None)
             if not node:
                 raise UserErrorException(f"Cannot find {node_name} in the flow.")
             for name, value in node.inputs.items():
                 if value.value_type == InputValueType.NODE_REFERENCE:
-                    input_name = f"{value.value}.{value.section}"
+                    input_name = (
+                        f"{value.value}.{value.section}.{value.property}"
+                        if value.property
+                        else f"{value.value}.{value.section}"
+                    )
                     if input_name in inputs:
                         dependency_input = inputs.pop(input_name)
                     elif name in inputs:
@@ -104,7 +114,9 @@ class TestSubmitter:
                     else:
                         missing_inputs.append(name)
                         continue
-                    dependency_nodes_outputs[value.value] = dependency_input
+                    dependency_nodes_outputs[value.value] = (
+                        {value.property: dependency_input} if value.property else dependency_input
+                    )
                     merged_inputs[name] = dependency_input
                 elif value.value_type == InputValueType.FLOW_INPUT:
                     input_name = f"{value.prefix}{value.value}"
@@ -113,7 +125,7 @@ class TestSubmitter:
                     elif name in inputs:
                         flow_input = inputs.pop(name)
                     else:
-                        flow_input = self.dataplane_flow.inputs[value.value].default
+                        flow_input = dataplane_flow.inputs[value.value].default
                         if flow_input is None:
                             missing_inputs.append(name)
                             continue
@@ -123,7 +135,7 @@ class TestSubmitter:
                     flow_inputs[name] = inputs.pop(name) if name in inputs else value.value
                     merged_inputs[name] = flow_inputs[name]
         else:
-            for name, value in self.dataplane_flow.inputs.items():
+            for name, value in dataplane_flow.inputs.items():
                 if name in inputs:
                     flow_inputs[name] = inputs.pop(name)
                     merged_inputs[name] = flow_inputs[name]
@@ -157,7 +169,7 @@ class TestSubmitter:
         stream_output: bool = True,
     ):
         from promptflow._constants import LINE_NUMBER_KEY
-        from promptflow.executor.flow_executor import FlowExecutor
+        from promptflow.executor import FlowExecutor
 
         if not connections:
             connections = SubmitterHelper.resolve_connections(flow=self.flow, client=self._client)
@@ -233,32 +245,16 @@ class TestSubmitter:
     def exec_with_inputs(self, inputs):
         # TODO: unify all exec_line calls here
         from promptflow._constants import LINE_NUMBER_KEY
-        from promptflow.executor.flow_executor import FlowExecutor
 
-        # validate connection objs
-        connection_obj_dict = {}
-        for key, connection_obj in self.flow_context.connection_objs.items():
-            scrubbed_secrets = connection_obj._get_scrubbed_secrets()
-            if scrubbed_secrets:
-                raise UserErrorException(
-                    f"Connection {connection_obj} contains scrubbed secrets with key {scrubbed_secrets.keys()}, "
-                    "please make sure connection has decrypted secrets to use in flow execution. "
-                )
-            connection_obj_dict[key] = connection_obj._to_execution_connection_dict()
-        connections = SubmitterHelper.resolve_connections(
-            flow=self.flow, client=self._client, connections_to_ignore=self.flow_context.connection_objs.keys()
-        )
-        # update connections with connection objs
-        connections.update(connection_obj_dict)
         # resolve environment variables
         SubmitterHelper.resolve_environment_variables(
             environment_variables=self.flow_context.environment_variables, client=self._client
         )
         SubmitterHelper.init_env(environment_variables=self.flow_context.environment_variables)
-        flow_executor = FlowExecutor.create(
-            flow_file=self.flow.path, connections=connections, working_dir=self.flow.code, raise_ex=True
-        )
-        flow_executor.enable_streaming_for_llm_flow(lambda: self.flow_context.streaming)
+        # cache executor here
+        flow_executor = FlowContextResolver.create(flow=self.flow)
+        # validate inputs
+        flow_inputs, _ = self.resolve_data(inputs=inputs, dataplane_flow=flow_executor._flow)
         line_result = flow_executor.exec_line(inputs, index=0, allow_generator_output=self.flow_context.streaming)
         if isinstance(line_result.output, dict):
             # Remove line_number from output
@@ -415,3 +411,72 @@ class TestSubmitter:
     def _get_generator_outputs(outputs):
         outputs = outputs or {}
         return {key: outputs for key, output in outputs.items() if isinstance(output, GeneratorType)}
+
+
+class TestSubmitterViaProxy(TestSubmitter):
+    def __init__(self, flow: Flow, flow_context: FlowContext, client=None):
+        super().__init__(flow, flow_context, client)
+
+    def flow_test(
+        self,
+        inputs: Mapping[str, Any],
+        environment_variables: dict = None,
+        stream_log: bool = True,
+        allow_generator_output: bool = False,
+        connections: dict = None,  # executable connections dict, to avoid http call each time in chat mode
+        stream_output: bool = True,
+    ):
+
+        from promptflow._constants import LINE_NUMBER_KEY
+        from promptflow.batch._csharp_executor_proxy import CsharpExecutorProxy
+
+        if not connections:
+            connections = SubmitterHelper.resolve_connection_names_from_tool_meta(
+                tools_meta=CsharpExecutorProxy.generate_tool_metadata(
+                    flow_dag=self.flow.dag,
+                    working_dir=self.flow.code,
+                )
+            )
+        credential_list = ConnectionManager(connections).get_secret_list()
+
+        # resolve environment variables
+        SubmitterHelper.resolve_environment_variables(environment_variables=environment_variables, client=self._client)
+        environment_variables = environment_variables if environment_variables else {}
+        SubmitterHelper.init_env(environment_variables=environment_variables)
+
+        with LoggerOperations(
+            file_path=self.flow.code / PROMPT_FLOW_DIR_NAME / "flow.log",
+            stream=stream_log,
+            credential_list=credential_list,
+        ):
+            storage = DefaultRunStorage(base_dir=self.flow.code, sub_dir=Path(".promptflow/intermediate"))
+            flow_executor = CsharpExecutorProxy.create(
+                flow_file=self.flow.path,
+                working_dir=self.flow.code,
+                connections=connections,
+                storage=storage,
+            )
+
+            line_result = flow_executor.exec_line(
+                inputs,
+                index=0,
+            )
+            line_result.output = persist_multimedia_data(
+                line_result.output, base_dir=self.flow.code, sub_dir=Path(".promptflow/output")
+            )
+            if line_result.aggregation_inputs:
+                # Convert inputs of aggregation to list type
+                flow_inputs = {k: [v] for k, v in inputs.items()}
+                aggregation_inputs = {k: [v] for k, v in line_result.aggregation_inputs.items()}
+                aggregation_results = flow_executor.exec_aggregation_async(
+                    flow_inputs, aggregation_inputs=aggregation_inputs
+                )
+                line_result.node_run_infos.update(aggregation_results.node_run_infos)
+                line_result.run_info.metrics = aggregation_results.metrics
+            if isinstance(line_result.output, dict):
+                # Remove line_number from output
+                line_result.output.pop(LINE_NUMBER_KEY, None)
+                generator_outputs = self._get_generator_outputs(line_result.output)
+                if generator_outputs:
+                    logger.info(f"Some streaming outputs in the result, {generator_outputs.keys()}")
+            return line_result